debug user avatar uploader

+++ server/bookHubServer/bookHubServer/Controllers/EditPersonalInfoController.cs
@@ -50,8 +50,11 @@ public class EditPersonalInfoController : ControllerBase
                     if(reader.Read())
                     {
                         string avartarUrl = reader.GetString("avatarImg");
-                        string serverLocalPath = @"/root/document/project/gitProject/hdu-bookhub/server/bookHubServer/bookHubServer/Assets/AvatarImg/" + avartarUrl.Substring(31);
-                        System.IO.File.Delete(serverLocalPath);
+                        if(avartarUrl != null)
+                        {
+                            string serverLocalPath = @"/root/document/project/gitProject/hdu-bookhub/server/bookHubServer/bookHubServer/Assets/AvatarImg/" + avartarUrl.Substring(31);
+                            System.IO.File.Delete(serverLocalPath);
+                        }
                     }
 
                 }
